system/requirements.txt
accelerate==0.30.1
aiohttp==3.9.5
aiosignal==1.3.1
anyio==4.4.0
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.2.0
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.30.2
backoff==2.2.1
beautifulsoup4==4.12.3
bitsandbytes==0.42.0
blis==0.7.11
boilerpy3==1.0.7
catalogue==2.0.10
cattrs==23.2.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
confection==0.1.5
contourpy==1.1.1
cycler==0.12.1
cymem==2.0.8
datasets==2.19.2
dill==0.3.8
docopt==0.6.2
el-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/el_core_news_sm-3.7.0/el_core_news_sm-3.7.0-py3-none-any.whl#sha256=8e061ef7dbf7dbeeae8d90a6d7c3457cc39f519dfcabfbe2b5138b654fd8c5bb
elastic-transport==7.16.0
elasticsearch==7.17.9
Events==0.5
exceptiongroup==1.2.1
farm-haystack==1.25.5
filelock==3.13.1
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.2
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
inflect==7.2.1
isodate==0.6.1
Jinja2==3.1.3
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langcodes==3.4.0
langdetect==1.0.9
language_data==1.2.0
lazy-imports==0.3.1
lxml==5.2.2
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
msrest==0.7.1
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
networkx==3.1
nltk==3.8.1
num2words==0.5.13
numpy==1.24.1
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.0
pandas==2.0.3
pillow==10.2.0
pkg_resources==0.0.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
posthog==3.5.0
preshed==3.0.9
prompthub-py==4.0.0
protobuf==5.27.1
psutil==5.9.8
pyarrow==16.1.0
pyarrow-hotfix==0.6
pydantic==1.10.15
Pygments==2.18.0
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-frontmatter==1.1.0
python-magic==0.4.27
python-pptx==0.6.23
pytz==2024.1
PyYAML==6.0.1
quantulum3==0.9.1
rank-bm25==0.2.2
referencing==0.35.1
regex==2024.5.15
requests==2.32.2
requests-cache==0.9.8
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.1
safetensors==0.4.3
scikit-learn==1.3.2
scipy==1.10.1
selenium==4.21.0
sentence-transformers==3.0.1
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sseclient-py==1.8.0
sympy==1.12
tenacity==8.3.0
thinc==8.2.4
threadpoolctl==3.5.0
tika==2.6.0
tiktoken==0.7.0
tokenizers==0.15.2
torch==2.3.0+cu118
torchaudio==2.3.0+cu118
torchvision==0.18.0+cu118
tqdm==4.66.4
transformers==4.39.3
trio==0.25.1
trio-websocket==0.11.1
triton==2.3.0
typeguard==4.3.0
typer==0.12.3
typing_extensions==4.10.0
tzdata==2024.1
url-normalize==1.4.3
urllib3==1.26.18
wasabi==1.1.3
weasel==0.4.1
wikiextractor==3.0.6
wrapt==1.16.0
wsproto==1.2.0
xformers==0.0.26.post1
XlsxWriter==3.2.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.0
